+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Fix global $woocommerce must be defined
 * WooCommerce: Remove snackbar notices
+* WooCommerce: Duplicate product changes the year if last year is present in the name or SKU 
 
 ### 1.3.8: May 3rd, 2021
 * WooCommerce: Fix Customize order completed and processing customer emails to protect local pickup and virtual informations